requirements.txt
# Web3 & Blockchain
web3==6.15.1
eth-account==0.11.0
eth-utils==4.1.0
gql==3.5.0
aiohttp==3.9.3

# API & Web
fastapi==0.109.2
uvicorn[standard]==0.27.1
streamlit==1.31.1
pydantic==2.6.1
python-multipart==0.0.9

# Image Processing
Pillow==10.2.0
requests==2.31.0

# IPFS
ipfshttpclient==0.8.0a2

# Data Processing
pandas==2.2.0
python-dotenv==1.0.1

# Utilities
httpx==0.27.0
